locust-load-test/requirements.txt
locust==2.17.0
kafka-python==2.0.2
requests==2.31.0